requirements/base.txt
# Django requirements
asgiref==3.4.1
Django==3.2.12
pytz==2021.3
sqlparse==0.4.2
requests==2.26.0
djangorestframework==3.13.1
django-mptt==0.14.0
django-mptt-admin==2.4.1

# DB requirements
psycopg2-binary==2.9.3

# # Image processing requirements
Pillow==8.4.0

# # Environments
python-dotenv==0.21.0
